refactor: streamline imports and remove unnecessary skips in test_global_metadata
