+++ tests/test_global_metadata.py
@@ -1,21 +1,12 @@
-"""Tests for OMEStream.set_global_metadata across backends."""
-
 from __future__ import annotations
 
 import json
-import threading
 from typing import TYPE_CHECKING, Any
 
 import numpy as np
 import pytest
 
-from ome_writers import (
-    AcquisitionSettings,
-    Dimension,
-    Plate,
-    Position,
-    create_stream,
-)
+from ome_writers import AcquisitionSettings, Dimension, Plate, Position, create_stream
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -47,8 +38,6 @@ def _settings(
 # ---------------------------------------------------------------------------
 # Zarr backend
 # ---------------------------------------------------------------------------
-
-pytest.importorskip("zarr", reason="zarr not available")
 
 
 def _zarr_attrs(path: Path) -> dict:
@@ -132,136 +121,3 @@ def test_global_metadata_invalid_namespace(tmp_path: Path, zarr_backend: str) ->
         with pytest.raises(ValueError, match="non-empty"):
             stream.set_global_metadata("", {})
         stream.append(FRAME)
-
-
-# ---------------------------------------------------------------------------
-# TIFF backend
-# ---------------------------------------------------------------------------
-
-try:
-    from ome_types import from_tiff, from_xml
-except ImportError:
-    pytest.skip("ome_types not installed", allow_module_level=True)
-
-
-def _anns_by_ns(ome_obj: Any, namespace: str) -> list[Any]:
-    sa = ome_obj.structured_annotations
-    if sa is None:
-        return []
-    return [a for a in sa.map_annotations if a.namespace == namespace]
-
-
-def _decode(annotation: Any) -> dict:
-    return {entry.k: json.loads(entry.value) for entry in annotation.value.ms}
-
-
-def test_tiff_global_metadata_single_file(tmp_path: Path, tiff_backend: str) -> None:
-    """Single-file TIFF: replace + siblings + not referenced by any plane."""
-    path = tmp_path / "single.ome.tiff"
-    fmt = {"name": "ome-tiff", "backend": tiff_backend}
-    dims = [
-        Dimension(name="t", count=2, type="time"),
-        Dimension(name="y", count=16, type="space"),
-        Dimension(name="x", count=16, type="space"),
-    ]
-    summary = {"mda": {"positions": 1}, "note": "single"}
-
-    with create_stream(_settings(path, fmt, dims)) as stream:
-        stream.set_global_metadata("ns_a", {"first": True})
-        stream.set_global_metadata("ns_a", summary)  # replace
-        stream.set_global_metadata("ns_b", {"y": 2})  # sibling
-        stream.append(FRAME)
-        stream.append(FRAME)
-
-    ome_obj = from_tiff(str(path))
-    ns_a = _anns_by_ns(ome_obj, "ns_a")
-    ns_b = _anns_by_ns(ome_obj, "ns_b")
-    assert len(ns_a) == 1 and _decode(ns_a[0]) == summary
-    assert len(ns_b) == 1 and _decode(ns_b[0]) == {"y": 2}
-
-    # Not referenced by any plane.
-    ann_ids = {ns_a[0].id, ns_b[0].id}
-    for image in ome_obj.images:
-        for plane in image.pixels.planes:
-            assert all(ref.id not in ann_ids for ref in plane.annotation_refs)
-
-
-@pytest.mark.parametrize("mode", ["companion-file", "master-tiff", "redundant"])
-def test_tiff_global_metadata_multi_file_modes(
-    tmp_path: Path, tiff_backend: str, mode: str
-) -> None:
-    """Each multi-file mode routes the annotation to the correct file(s)."""
-    d = tmp_path / mode
-    fmt = {"name": "ome-tiff", "backend": tiff_backend, "multi_file_metadata": mode}
-    dims = [
-        Dimension(name="p", type="position", coords=["Pos0", "Pos1"]),
-        Dimension(name="t", count=2, type="time"),
-        Dimension(name="y", count=16, type="space"),
-        Dimension(name="x", count=16, type="space"),
-    ]
-    summary = {"mda": mode}
-
-    with create_stream(_settings(d, fmt, dims)) as stream:
-        stream.set_global_metadata("ns", summary)
-        for _ in range(4):
-            stream.append(FRAME)
-
-    per_pos = [
-        _anns_by_ns(from_tiff(str(d / f"{mode}_p{p:03d}.ome.tiff")), "ns")
-        for p in range(2)
-    ]
-    if mode == "redundant":
-        for anns in per_pos:
-            assert len(anns) == 1 and _decode(anns[0]) == summary
-    elif mode == "master-tiff":
-        assert len(per_pos[0]) == 1 and _decode(per_pos[0][0]) == summary
-        assert per_pos[1] == []
-    else:  # companion-file
-        assert per_pos[0] == [] and per_pos[1] == []
-        companion = from_xml((d / "companion.ome").read_text())
-        c_anns = _anns_by_ns(companion, "ns")
-        assert len(c_anns) == 1 and _decode(c_anns[0]) == summary
-
-
-def test_tiff_global_metadata_concurrent_with_close(
-    tmp_path: Path, tiff_backend: str
-) -> None:
-    """Concurrent summary updates and close should not crash."""
-    path = tmp_path / "concurrent_close.ome.tiff"
-    fmt = {"name": "ome-tiff", "backend": tiff_backend}
-    dims = [
-        Dimension(name="t", count=16, type="time"),
-        Dimension(name="y", count=16, type="space"),
-        Dimension(name="x", count=16, type="space"),
-    ]
-    errors: list[BaseException] = []
-    stop = threading.Event()
-
-    stream = create_stream(_settings(path, fmt, dims))
-    stream.set_global_metadata("ns", {"i": 0})
-
-    def _loop() -> None:
-        i = 1
-        while not stop.is_set():
-            try:
-                stream.set_global_metadata("ns", {"i": i})
-            except RuntimeError:  # expected once close finalizes
-                break
-            except BaseException as e:
-                errors.append(e)
-                break
-            i += 1
-
-    worker = threading.Thread(target=_loop)
-    worker.start()
-    try:
-        for _ in range(8):
-            stream.append(FRAME)
-    finally:
-        stream.close()
-        stop.set()
-        worker.join(timeout=2)
-
-    assert not errors
-    matches = _anns_by_ns(from_tiff(str(path)), "ns")
-    assert len(matches) == 1
